+++ coinbase/order.js
@@ -3,10 +3,12 @@
 const { divide, multiply } = require('../lib/math');
 const config = require('../config');
 const memory = require('../lib/memory');
+const log = require('../lib/log');
 const numFix = require('../lib/number.fix');
 const request = require('./cb.request');
+const sleep = require('../lib/sleep');
 
-module.exports = async opts => {
+module.exports = async (opts, retries = 0) => {
   if (config.dry) {
     // fake out a .5% fee subtraction (worst case tier)
     const converted = multiply(opts.funds, 0.995);
@@ -23,10 +25,36 @@ module.exports = async opts => {
       settled: true,
     };
   }
-  const response = await request({
-    requestPath: '/orders',
-    method: 'POST',
-    body: opts,
-  });
-  return response ? response.json : response;
+
+  let retryCount = 0;
+  let time = 100;
+  const attemptOrder = async () => {
+    if (retryCount) {
+      // if we had something like an ENETDOWN error, try again after a bit
+      time = 5000; // extend future retries to 5 seconds
+    }
+    if (retryCount === 2) {
+      log.now(`retry #${retryCount}`, opts);
+    }
+    const response = await request({
+      requestPath: '/orders',
+      method: 'POST',
+      body: opts,
+    });
+    log.debug(response);
+    const json = response ? response.json : response;
+    // if retries are enabled for this type of order, allow retry
+    if (retries && !json) {
+      retryCount++;
+      if (retryCount > retries) {
+        log.error(`failed to place order`, { opts, json });
+        return false;
+      }
+      await sleep(time);
+      return attemptOrder();
+    }
+    return json;
+  };
+
+  return attemptOrder();
 };